+++ Samples/HTMLUI/QuasarBasedHTMLApp/src/router/index.ts
@@ -1,8 +1,6 @@
 import { route } from 'quasar/wrappers';
 import {
-  createMemoryHistory,
   createRouter,
-  createWebHashHistory,
   createWebHistory,
 } from 'vue-router';
 
@@ -18,21 +16,10 @@ import routes from './routes';
  */
 
 export default route(function (/* { store, ssrContext } */) {
-  const createHistory = process.env.SERVER
-    ? createMemoryHistory
-    : process.env.VUE_ROUTER_MODE === 'history'
-    ? createWebHistory
-    : createWebHashHistory;
-
   const Router = createRouter({
     scrollBehavior: () => ({ left: 0, top: 0 }),
     routes,
-
-    // Leave this as is and make changes in quasar.conf.js instead!
-    // quasar.conf.js -> build -> vueRouterMode
-    // quasar.conf.js -> build -> publicPath
-    history: createHistory(process.env.VUE_ROUTER_BASE),
+    history: createWebHistory(),  // because google OAUTH doesn't allow fragments(hash) in callback url,
   });
-
   return Router;
 });